+++ reikna/cluda/cuda.py
@@ -13,7 +13,7 @@ import reikna.cluda as cluda
 import reikna.cluda.dtypes as dtypes
 from reikna.helpers import factors, wrap_in_tuple, product
 from reikna.cluda.kernel import render_prelude, render_template_source
-from reikna.cluda.vsize import VirtualSizes, render_stub_vsize_funcs
+from reikna.cluda.vsize import VirtualSizes
 from reikna.cluda.tempalloc import ZeroOffsetManager
 
 
@@ -196,9 +196,9 @@ class Context:
         return Module(self, template_src, render_kwds=render_kwds)
 
     def compile_static(self, template_src, name, global_size,
-            local_size=None, local_mem=0, render_kwds=None):
+            local_size=None, local_mem=0, render_args=None, render_kwds=None):
         return StaticKernel(self, template_src, name, global_size,
-            local_size=local_size, render_kwds=render_kwds)
+            local_size=local_size, render_args=render_args, render_kwds=render_kwds)
 
     def _pytest_finalize(self):
         """
@@ -323,18 +323,25 @@ class Kernel:
 
 class StaticKernel:
 
-    def __init__(self, ctx, src, name, global_size, local_size=None, render_kwds=None):
+    def __init__(self, ctx, template_src, name, global_size, local_size=None,
+            render_args=None, render_kwds=None):
+
         self._ctx = ctx
 
+        if render_args is None:
+            render_args = []
         if render_kwds is None:
             render_kwds = {}
 
         prelude = render_prelude(self._ctx)
-        stub_vsize_funcs = render_stub_vsize_funcs()
-        src = render_template_source(src, **render_kwds)
+        src = render_template_source(template_src, *render_args, **render_kwds)
 
         # We need the first approximation of the maximum thread number for a kernel.
         # Stub virtual size functions instead of real ones will not change it (hopefully).
+        stub_vs = VirtualSizes(ctx.device_params, ctx.device_params.max_work_group_size,
+            global_size, local_size)
+        stub_vsize_funcs = stub_vs.render_vsize_funcs()
+
         stub_module = ctx._compile(str(prelude + stub_vsize_funcs + src))
         stub_kernel = stub_module.get_function(name)
         max_work_group_size = stub_kernel.get_attribute(