+++ reikna/cluda/dtypes.py
@@ -127,13 +127,15 @@ def c_constant(val, dtype=None):
     """
     if dtype is None:
         dtype = detect_type(val)
+    val = numpy.cast[dtype](val)
+
     if is_complex(dtype):
         return "COMPLEX_CTR(" + ctype(dtype) + ")(" + \
             c_constant(val.real) + ", " + c_constant(val.imag) + ")"
     elif is_integer(dtype):
         return str(val) + ("L" if dtype.itemsize > 4 else "")
     else:
-        return repr(val) + ("f" if dtype.itemsize <= 4 else "")
+        return repr(float(val)) + ("f" if dtype.itemsize <= 4 else "")
 
 def _register_dtype(dtype, ctype):
     dtype = normalize_type(dtype)