+++ doc/source/history.rst
@@ -2,27 +2,82 @@
 Release history
 ***************
 
-0.3.0 (current development version)
+
+0.2.3 (current development version)
 ===================================
 
+Under construction.
+
+
+0.2.2 (20 Apr 2013)
+===================
+
+* ADDED: ability to use custom argument names in transformations.
+
+* ADDED: multi-argument :py:func:`~reikna.cluda.functions.mul`.
+
+* ADDED: counter-based random number generator :py:class:`~reikna.cbrng.CBRNG`.
+
+* ADDED: :py:class:`~reikna.elementwise.Elementwise` now supports argument dependencies.
+
+* ADDED: Module support in CLUDA; see :ref:`tutorial-modules` for details.
+
+* ADDED: :py:func:`~reikna.helpers.template_func`.
+
+* CHANGED: :py:func:`~reikna.cluda.kernel.render_template_source` is the main renderer now.
+
+* CHANGED: ``FuncCollector`` class was removed; functions are now used as common modules.
+
+* CHANGED: all templates created with :py:func:`~reikna.helpers.template_for` are now rendered with ``from __future__ import division``.
+
+* CHANGED: signature of :py:meth:`~reikna.core.operation.OperationRecorder.add_kernel` takes a renderable instead of a full template.
+
+* CHANGED: :py:meth:`~reikna.cluda.api.Context.compile_static` now takes a template instead of a source.
+
+* CHANGED: :py:class:`~reikna.elementwise.Elementwise` now uses modules.
+
+* FIXED: potential problem with local size finidng in static kernels (first approximation for the maximum workgroup size was not that good)
+
+* FIXED: some OpenCL compilation warnings caused by an incorrect version querying macro.
+
+* FIXED: bug with incorrect processing of scalar global size in static kernels.
+
+* FIXED: bug in variance estimates in CBRNG tests.
+
+* FIXED: error in the temporary varaiable type in :py:func:`reikna.cluda.functions.polar` and :py:func:`reikna.cluda.functions.exp`.
+
+
+0.2.1 (8 Mar 2013)
+==================
+
 * FIXED: function names for kernel ``polar()``, ``exp()`` and ``conj()``.
+
 * FIXED: added forgotten kernel ``norm()`` handler.
+
 * FIXED: bug in ``Py.Test`` testcase execution hook which caused every test to run twice.
+
 * FIXED: bug in nested computation processing for computation with more than one kernel.
-* FIXED: added dependencies between MatrixMul kernel arguments.
+
+* FIXED: added dependencies between :py:class:`~reikna.matrixmul.MatrixMul` kernel arguments.
+
 * FIXED: taking into account dependencies between input and output arrays as well as the ones
-    between internal allocations --- necessary for nested computations.
-* ADDED: discrete harmonic transform :py:class:`reikna.dht.DHT`
-    (calculated using Gauss-Hermite quadrature).
+  between internal allocations --- necessary for nested computations.
+
+* ADDED: discrete harmonic transform :py:class:`~reikna.dht.DHT`
+  (calculated using Gauss-Hermite quadrature).
 
 
 0.2.0 (3 Mar 2013)
 ==================
 
 * Added FFT computation (slightly optimized PyFFT version + Bluestein's algorithm for non-power-of-2 FFT sizes)
+
 * Added Python 3 compatibility
+
 * Added Context-global automatic memory packing
+
 * Added polar(), conj() and exp() functions to kernel toolbox
+
 * Changed name because of the clash with `another Tigger <http://www.astron.nl/meqwiki/Tigger>`_.
 
 
@@ -30,7 +85,9 @@ Release history
 ===================
 
 * Lots of changes in the API
+
 * Added elementwise, reduction and transposition computations
+
 * Extended API reference and added topical guides
 
 
@@ -38,5 +95,7 @@ Release history
 ===================
 
 * Created basic core for computations and transformations
+
 * Added matrix multiplication computation
+
 * Created basic documentation